Create function for base retrieval of access token
Moved the webserver code into this function. When the function is
executed, the server is turned on, the url is opened in the browser
to initiate the authentication flow, and the webserver is closed when
the access token is returned.
The closing of the window is handled by a script in the html window.